refactor(hapi): split producer service
