+++ hapi/src/services/producer.service.js
@@ -1,8 +1,9 @@
-const { hasuraUtil, sequelizeUtil, producerUtil } = require('../utils')
+const { hasuraUtil, sequelizeUtil } = require('../utils')
 const { eosConfig } = require('../config')
 
 const lacchainService = require('./lacchain.service')
 const eosioService = require('./eosio.service')
+const NodeService = require('./node.service')
 
 const updateProducers = async (producers = []) => {
   const upsertMutation = `
@@ -32,55 +33,10 @@ const updateProducers = async (producers = []) => {
   return insertedRows.insert_producer.returning
 }
 
-const clearNodes = async () => {
-  const clearMutation = `
-    mutation{
-      delete_node(where: {}){
-        affected_rows
-      }
-    }
-  `
-
-  await hasuraUtil.request(clearMutation)
-}
-
-const updateNodes = async (nodes = []) => {
-  await clearNodes()
-
-  const upsertMutation = `
-    mutation ($nodes: [node_insert_input!]!) {
-      insert_node(objects: $nodes, on_conflict: {constraint: node_pkey,update_columns: [type,full,location,producer_id]}) {
-        affected_rows
-      }
-    }
-  `
-
-  await hasuraUtil.request(upsertMutation, { nodes })
-}
-
-const updateEndpointInfo = async (endpoint) => {
-  if (!endpoint.type || !['api', 'ssl'].includes(endpoint.type)) return
-
-  const updateMutation = `
-    mutation ($id: uuid, $head_block_num: Int) {
-      update_endpoint(_set: {head_block_num: $head_block_num}, where: {id: {_eq: $id}}) {
-        affected_rows
-      }
-    }
-  `
-
-  const endpointInfo = await producerUtil.getNodeInfo(endpoint.value)
-
-  await hasuraUtil.request(updateMutation, {
-    id: endpoint.id,
-    head_block_num: endpointInfo.head_block_num || 0
-  })
-}
-
 const syncProducers = async () => {
   let producers = []
 
-  await clearNodes()
+  await NodeService.clearNodes()
 
   switch (eosConfig.networkName) {
     case eosConfig.knownNetworks.lacchain:
@@ -93,6 +49,7 @@ const syncProducers = async () => {
 
   producers = await updateProducers(producers)
   await syncNodes(producers)
+  await syncEndpoints()
 }
 
 const getProducersSummary = async () => {
@@ -115,45 +72,13 @@ const syncNodes = async (producers) => {
 
   const nodes = producers.flatMap((producer) => {
     return (producer.bp_json?.nodes || []).map((node) => {
-      const endpoints = []
-      const types = ['api', 'ssl', 'p2p']
+      node.producer_id = producer.id
 
-      types.forEach((type) => {
-        if (node[type + '_endpoint']) {
-          endpoints.push({
-            type: type,
-            value: node[type + '_endpoint']
-          })
-        }
-      })
-
-      let formatNode = {
-        type: Array.isArray(node.node_type) ? node.node_type : [node.node_type],
-        full: node.full ?? false,
-        location: node.location ?? {},
-        producer_id: producer.id
-      }
-
-      if (endpoints.length) {
-        formatNode.endpoints = {
-          data: endpoints
-        }
-      }
-
-      if (node.features?.length) {
-        formatNode.node_info = {
-          data: {
-            version: node.server_version_string ?? '',
-            features: node.features
-          }
-        }
-      }
-
-      return formatNode
+      return NodeService.getFormatNode(node)
     })
   })
 
-  await updateNodes(nodes)
+  await NodeService.updateNodes(nodes)
 }
 
 const syncEndpoints = async () => {
@@ -169,7 +94,7 @@ const syncEndpoints = async () => {
   if (!endpoints?.length) return
 
   endpoints.forEach(async (endpoint) => {
-    await updateEndpointInfo(endpoint)
+    await NodeService.updateEndpointInfo(endpoint)
   })
 }
